chore(shared): centralize activity types
