+++ IMPLEMENTATION_PLAN.md
@@ -3,10 +3,10 @@
 ## 1. Type System Unification
 
 ### 1.1 Shared Types Migration
-- [ ] Create a dedicated `@shared/types` package
-- [ ] Migrate all type definitions to the shared package
-- [ ] Set up proper versioning for the shared package
-- [ ] Update both client and server to consume the shared types
+- [x] Create a dedicated `@shared/types` package
+- [x] Migrate all type definitions to the shared package
+- [x] Set up proper versioning for the shared package
+- [x] Update both client and server to consume the shared types
 
 ### 1.2 Type Safety Enhancements
 - [ ] Enable TypeScript's `strict` mode
