+++ audit-report.md
@@ -777,7 +777,7 @@ The database uses PostgreSQL with Drizzle ORM for schema definition and query bu
 
 ### 1.1 Shared Types Migration (2-3 weeks)
 - [x] Create a dedicated `@shared/types` package
-- [ ] Migrate all type definitions to the shared package
+- [x] Migrate all type definitions to the shared package
 - [x] Set up proper versioning for the shared package
 - [x] Update both client and server to consume the shared types
