+++ shared/types/index.ts
@@ -11,3 +11,11 @@ export type { RegisterDto } from './auth/dto/register.dto.js';
 export type { RequestPasswordResetDto, ResetPasswordDto, ChangePasswordDto } from './auth/dto/password-reset.dto.js';
 export type { UserResponse } from "./auth/dto/user-response.dto.js";
 export { AppErrorCode } from './error-codes.js';
+// Export common domain types
+export * from './api.js';
+export * from './trips.js';
+export * from './bookings.js';
+export * from './invoice.js';
+export * from './users.js';
+export * from './organizations.js';
+export * from './activity.js';
